requirements.txt
tensorflow>=2.10.0
numpy>=1.21.0
matplotlib>=3.5.0
scikit-learn>=1.1.0
pillow>=9.0.0
seaborn>=0.11.0
mlflow>=2.0.0
pathlib2>=2.3.7; python_version < "3.4"
typing-extensions>=4.0.0; python_version < "3.8"
fastapi>=0.109.0
uvicorn>=0.22.0
python-multipart>=0.0.9
Pillow>=9.0.0
